Backed out changeset 14a41aec9db8 (bug 1687400) for Valgrind failure. CLOSED TREE
UltraBlame original commit: 0ac6c7e669a83788a44d08cb255383bd83621a38

+++ js/src/vm/Xdr.h
@@ -830,7 +830,7 @@ class XDRIncrementalStencilEncoder : public XDRIncrementalEncoderBase {
   
   
   
-  using FunctionKey = uint32_t;
+  using FunctionKey = uint64_t;
   HashSet<FunctionKey> encodedFunctions_;
 
  public:
